refactor : move RecoilObserver component to test-utils folder
