+++ src/utils/test-utils/index.ts
@@ -1,2 +1,3 @@
 export * from './mockRouter';
 export * from './mockUseSession';
+export * from './RecoilObserver';
